+++ src/Dapperer/QueryBuilders/MsSql/StringTableValuedParams.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.SqlServer.Server;
 
-namespace Dapperer
+namespace Dapperer.QueryBuilders.MsSql
 {
     public class StringTableValuedParams : TableValuedParams
     {
