+++ src/Tests/Dapperer.Tests.Unit/SqlQueryBuilderTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Dapperer.QueryBuilders;
+using Dapperer.QueryBuilders.MsSql;
 using NUnit.Framework;
 
 namespace Dapperer.Tests.Unit
